parser-service/requirements.txt
pdf2image>=1.16.3
pytesseract>=0.3.10
python-docx>=1.0.1
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
PyPDF2>=3.0.0
nltk>=3.8.1
langdetect>=1.0.9
pydantic>=2.4.2
python-dateutil>=2.8.2
sentence-transformers>=2.2.2
torch>=2.0.0
redis>=5.0.0
